% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgreadlength.R
\name{avgreadlength}
\alias{avgreadlength}
\title{avgreadlength}
\usage{
avgreadlength(x, y)
}
\arguments{
\item{x}{is a bam file}

\item{y}{is a bam.bai file}
}
\value{
the average read length
}
\description{
function calculates the average read length
}
