% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_gata.R
\name{count_gata}
\alias{count_gata}
\title{count_gata}
\usage{
count_gata(data)
}
\arguments{
\item{data}{is a fasta file}
}
\value{
the number of "GATA" sequences in a fasta file
}
\description{
function calculates the number of "GATA" sequences in a fasta file
}
