% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapped.R
\name{mapped}
\alias{mapped}
\title{mapped}
\usage{
mapped(x, y)
}
\arguments{
\item{x}{is a bam file}

\item{y}{is a bam.bai file}
}
\value{
the number of mapped reads
}
\description{
function calculates the number of mapped reads
}
