% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvsnot.R
\name{mapvsnot}
\alias{mapvsnot}
\title{mapvsnot}
\usage{
mapvsnot(x, y)
}
\arguments{
\item{x}{is a bam file}

\item{y}{is a bam.bai file}
}
\value{
the total number of reads (mapped and unmapped)
}
\description{
function calculates the total number of reads (mapped and unmapped)
}
