% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentA.R
\name{percentA}
\alias{percentA}
\title{percentA}
\usage{
percentA(x)
}
\arguments{
\item{x}{is a fasta file}
}
\value{
the percentage of the base A in a fasta file
}
\description{
function calculates the percentage of the base A in a fasta file
}
