% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentC.R
\name{percentC}
\alias{percentC}
\title{percentC}
\usage{
percentC(x)
}
\arguments{
\item{x}{is a fasta file}
}
\value{
the percentage of the base C in a fasta file
}
\description{
function calculates the percentage of the base C in a fasta file
}
