% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentG.R
\name{percentG}
\alias{percentG}
\title{percentG}
\usage{
percentG(x)
}
\arguments{
\item{x}{is a fasta file}
}
\value{
the percentage of the base G in a fasta file
}
\description{
function calculates the percentage of the base G in a fasta file
}
