% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentmapped.R
\name{percentmapped}
\alias{percentmapped}
\title{percentmapped}
\usage{
percentmapped(x, y)
}
\arguments{
\item{x}{is a bam file}

\item{y}{is a bam.bai file}
}
\value{
the percentage  of reads that were mapped to the genome
}
\description{
function calculates the percentage  of mapped reads
}
