% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentT.R
\name{percentT}
\alias{percentT}
\title{percentT}
\usage{
percentT(x)
}
\arguments{
\item{x}{is a fasta file}
}
\value{
the percentage of the base T in a fasta file
}
\description{
function calculates the percentage of the base T in a fasta file
}
