% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentunmapped.R
\name{percentunmapped}
\alias{percentunmapped}
\title{percentunmapped}
\usage{
percentunmapped(x, y)
}
\arguments{
\item{x}{is a bam file}

\item{y}{is a bam.bai file}
}
\value{
the percentage  of reads that were NOT mapped to the genome
}
\description{
function calculates the percentage  of unmapped reads
}
