% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped.R
\name{unmapped}
\alias{unmapped}
\title{percentunmapped}
\usage{
unmapped(x, y)
}
\arguments{
\item{x}{is a bam file}

\item{y}{is a bam.bai file}
}
\value{
the number  of reads that were NOT mapped to the genome
}
\description{
function displays the number  of unmapped reads
}
